adapt example, use i18n on global not found page
