+++ examples/example-app-router/src/app/[locale]/layout.tsx
@@ -1,16 +1,8 @@
-import clsx from 'clsx';
-import {Inter} from 'next/font/google';
-import {NextIntlClientProvider} from 'next-intl';
-import {
-  getMessages,
-  getTranslations,
-  unstable_setRequestLocale
-} from 'next-intl/server';
+import {notFound} from 'next/navigation';
+import {getTranslations, unstable_setRequestLocale} from 'next-intl/server';
 import {ReactNode} from 'react';
-import Navigation from '@/components/Navigation';
+import BaseLayout from '@/components/BaseLayout';
 import {routing} from '@/i18n/routing';
-
-const inter = Inter({subsets: ['latin']});
 
 type Props = {
   children: ReactNode;
@@ -35,21 +27,13 @@ export default async function LocaleLayout({
   children,
   params: {locale}
 }: Props) {
+  // Abort render if an unknown value is received as the `[locale]` param
+  if (!routing.locales.includes(locale as any)) {
+    notFound();
+  }
+
   // Enable static rendering
   unstable_setRequestLocale(locale);
 
-  // Providing all messages to the client
-  // side is the easiest way to get started
-  const messages = await getMessages();
-
-  return (
-    <html className="h-full" lang={locale}>
-      <body className={clsx(inter.className, 'flex h-full flex-col')}>
-        <NextIntlClientProvider messages={messages}>
-          <Navigation />
-          {children}
-        </NextIntlClientProvider>
-      </body>
-    </html>
-  );
+  return <BaseLayout locale={locale}>{children}</BaseLayout>;
 }
