+++ examples/example-app-router/src/components/NotFoundPage.tsx
@@ -1,0 +1,12 @@
+import {useTranslations} from 'next-intl';
+import PageLayout from './PageLayout';
+
+export default function NotFoundPage() {
+  const t = useTranslations('NotFoundPage');
+
+  return (
+    <PageLayout title={t('title')}>
+      <p className="max-w-[460px]">{t('description')}</p>
+    </PageLayout>
+  );
+}
