+++ examples/example-app-router/src/i18n/request.ts
@@ -1,12 +1,16 @@
-import {notFound} from 'next/navigation';
 import {getRequestConfig} from 'next-intl/server';
 import {routing} from './routing';
 
-export default getRequestConfig(async ({locale}) => {
-  // Validate that the incoming `locale` parameter is valid
-  if (!routing.locales.includes(locale as any)) notFound();
+export default getRequestConfig(async ({requestLocale}) => {
+  let locale = await requestLocale;
+
+  // Ensure the locale is valid
+  if (!routing.locales.includes(locale as any)) {
+    locale = routing.defaultLocale;
+  }
 
   return {
+    locale,
     messages: (
       await (locale === 'en'
         ? // When using Turbopack, this will enable HMR for `en`
