+++ examples/example-app-router/tests/main.spec.ts
@@ -28,9 +28,10 @@ it("handles not found pages for routes that don't match the middleware", async (
   page
 }) => {
   await page.goto('/test.png');
-  page.getByRole('heading', {name: 'This page could not be found.'});
+  page.getByRole('heading', {name: 'Page not found'});
+
   await page.goto('/api/hello');
-  page.getByRole('heading', {name: 'This page could not be found.'});
+  page.getByRole('heading', {name: 'Page not found'});
 });
 
 it('sets caching headers', async ({request}) => {
